requirements.txt
cryptography>=41.0.0
pyOpenSSL>=23.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-jose>=3.3.0
python-json-logger>=2.0.7
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0
flask>=2.0.1
flask-socketio>=5.1.1
python-socketio>=5.4.0
